fix(pg): rm underscore, load pg-user secret in pg
